+++ changelog.md
@@ -2,9 +2,18 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9410]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9411]: https://www.github.com/atecoder/dn.vi.ivi
 
 Current IVI Visa Compatibility: 8.0.1.9391
+
+## [8.1.9411] - 2025-10-07
+- Update the versions to 8.1.9411.108
+- Device.TSP
+  - Session base extensions: 
+    - Add IsScriptExist method checking if the script is embedded or is a user script.
+    - Add IsScriptDeleted method.  
+	- Validate that script was deleted.
+  - Remove Ignore Existing from the script import and load methods.
 
 ## [8.1.9410] - 2025-10-06
 - Increment TTM firmware release version to 9410.